move libs to external/ and fix lint
