+++ .golangci.yaml
@@ -11,3 +11,8 @@ linters:
         - -QF1008 # https://staticcheck.dev/docs/checks/#QF1008 Omit embedded fields from selector expression.
         - -ST1000 # at least one file in a package should have a package comment
         - -ST1003 # struct field Api should be API
+  exclusions:
+    rules:
+      - path: external\/go4vl
+        linters:
+          - staticcheck
