+++ external/go4vl/v4l2/ext_controls.go
@@ -30,7 +30,7 @@ func GetExtControlValue(fd uintptr, ctrlID CtrlID) (CtrlValue, error) {
 func SetExtControlValue(fd uintptr, id CtrlID, val CtrlValue) error {
 	ctrlInfo, err := QueryExtControlInfo(fd, id)
 	if err != nil {
-		return fmt.Errorf("set ext control value: id %s: %w", id, err)
+		return fmt.Errorf("set ext control value: id %d: %w", id, err)
 	}
 	if val < ctrlInfo.Minimum || val > ctrlInfo.Maximum {
 		return fmt.Errorf("set ext control value: out-of-range failure: val %d: expected ctrl.Min %d, ctrl.Max %d", val, ctrlInfo.Minimum, ctrlInfo.Maximum)
@@ -87,7 +87,7 @@ func GetExtControl(fd uintptr, id CtrlID) (Control, error) {
 	// retrieve control value
 	ctrlValue, err := GetExtControlValue(fd, uint32(id))
 	if err != nil {
-		return Control{}, fmt.Errorf("get control: %w", id, err)
+		return Control{}, fmt.Errorf("get control on %d: %w", id, err)
 	}
 
 	control.Value = ctrlValue
